refactor: agentic adapters → llm/providers/ 통합
agent/adapters/{claude,openai,glm}_agentic_adapter.py를 각
llm/providers/{anthropic,openai,glm}.py에 통합.
resolve_agentic_adapter + CROSS_PROVIDER_FALLBACK를 llm/router.py로 이동.
core/agent/adapters/ 디렉토리(5파일) 삭제. 모듈 195 → 187.
행동 변경 없음.
