+++ CLAUDE.md
@@ -8,7 +8,7 @@ LangGraph 기반 범용 자율 실행 에이전트. 리서치, 분석, 자동화
 - **Python**: >= 3.12
 - **Package Manager**: uv
 - **Entry Point**: `geode.cli:app` (Typer)
-- **Modules**: 195
+- **Modules**: 187
 - **Tests**: 3249+
 - **CHANGELOG**: `CHANGELOG.md` (Keep a Changelog + SemVer)
