--- core/agent/adapters/__init__.py
@@ -1,1 +1,0 @@
-"""Agentic LLM adapters — provider-specific implementations for AgenticLoop."""
