+++ tests/test_native_tools.py
@@ -55,7 +55,7 @@ class TestAnthropicNativeToolInjection:
 
     def test_native_tools_constant_has_web_search(self) -> None:
         """_ANTHROPIC_NATIVE_TOOLS must include web_search_20260209."""
-        from core.agent.adapters.claude_agentic_adapter import (
+        from core.llm.providers.anthropic import (
             _ANTHROPIC_NATIVE_TOOLS,
         )
 
@@ -66,7 +66,7 @@ class TestAnthropicNativeToolInjection:
 
     def test_native_tools_constant_has_web_fetch(self) -> None:
         """_ANTHROPIC_NATIVE_TOOLS must include web_fetch_20260209."""
-        from core.agent.adapters.claude_agentic_adapter import (
+        from core.llm.providers.anthropic import (
             _ANTHROPIC_NATIVE_TOOLS,
         )
 
@@ -77,7 +77,7 @@ class TestAnthropicNativeToolInjection:
 
     def test_native_tools_deduplication(self) -> None:
         """Native tools should not duplicate existing tool names."""
-        from core.agent.adapters.claude_agentic_adapter import (
+        from core.llm.providers.anthropic import (
             _ANTHROPIC_NATIVE_TOOLS,
             _API_ALLOWED_KEYS,
         )
@@ -114,7 +114,7 @@ class TestAnthropicNativeToolInjection:
         """agentic_call must include native tools in the API request."""
         import asyncio
 
-        from core.agent.adapters.claude_agentic_adapter import (
+        from core.llm.providers.anthropic import (
             ClaudeAgenticAdapter,
         )
 
@@ -127,8 +127,8 @@ class TestAnthropicNativeToolInjection:
         mock_response.model = "claude-opus-4-6"
 
         with (
-            patch("core.agent.adapters.claude_agentic_adapter.settings") as mock_settings,
-            patch("core.agent.adapters.claude_agentic_adapter.call_with_failover") as mock_failover,
+            patch("core.llm.providers.anthropic.settings") as mock_settings,
+            patch("core.llm.router.call_with_failover") as mock_failover,
         ):
             mock_settings.anthropic_api_key = "test-key"
             mock_failover.return_value = (mock_response, "claude-opus-4-6")
@@ -175,7 +175,7 @@ class TestGlmNativeWebSearch:
 
     def test_glm_native_web_search_constant(self) -> None:
         """_GLM_NATIVE_WEB_SEARCH must have correct structure."""
-        from core.agent.adapters.glm_agentic_adapter import (
+        from core.llm.providers.glm import (
             _GLM_NATIVE_WEB_SEARCH,
         )
 
@@ -184,10 +184,10 @@ class TestGlmNativeWebSearch:
 
     def test_glm_adapter_is_subclass_of_openai(self) -> None:
         """GlmAgenticAdapter must inherit from OpenAIAgenticAdapter."""
-        from core.agent.adapters.glm_agentic_adapter import (
+        from core.llm.providers.glm import (
             GlmAgenticAdapter,
         )
-        from core.agent.adapters.openai_agentic_adapter import (
+        from core.llm.providers.openai import (
             OpenAIAgenticAdapter,
         )
 
@@ -195,7 +195,7 @@ class TestGlmNativeWebSearch:
 
     def test_glm_provider_name(self) -> None:
         """GlmAgenticAdapter.provider_name must be 'glm'."""
-        from core.agent.adapters.glm_agentic_adapter import (
+        from core.llm.providers.glm import (
             GlmAgenticAdapter,
         )
 
@@ -204,10 +204,10 @@ class TestGlmNativeWebSearch:
 
     def test_glm_has_own_agentic_call(self) -> None:
         """GlmAgenticAdapter must override agentic_call (not inherit from OpenAI)."""
-        from core.agent.adapters.glm_agentic_adapter import (
+        from core.llm.providers.glm import (
             GlmAgenticAdapter,
         )
-        from core.agent.adapters.openai_agentic_adapter import (
+        from core.llm.providers.openai import (
             OpenAIAgenticAdapter,
         )
 
@@ -227,7 +227,7 @@ class TestOpenAIResponsesApiGap:
 
     def test_openai_adapter_has_todo_comment(self) -> None:
         """OpenAIAgenticAdapter docstring must mention Responses API TODO."""
-        from core.agent.adapters.openai_agentic_adapter import (
+        from core.llm.providers.openai import (
             OpenAIAgenticAdapter,
         )
 
@@ -239,7 +239,7 @@ class TestOpenAIResponsesApiGap:
         """Confirm OpenAI adapter still uses Chat Completions (not yet migrated)."""
         import inspect
 
-        from core.agent.adapters.openai_agentic_adapter import (
+        from core.llm.providers.openai import (
             OpenAIAgenticAdapter,
         )
 
@@ -257,7 +257,7 @@ class TestApiAllowedKeys:
 
     def test_type_key_in_allowed(self) -> None:
         """'type' must be in _API_ALLOWED_KEYS for native tool passthrough."""
-        from core.agent.adapters.claude_agentic_adapter import (
+        from core.llm.providers.anthropic import (
             _API_ALLOWED_KEYS,
         )
 
@@ -265,7 +265,7 @@ class TestApiAllowedKeys:
 
     def test_native_tool_passes_filter(self) -> None:
         """Native tool dict must survive _API_ALLOWED_KEYS filtering."""
-        from core.agent.adapters.claude_agentic_adapter import (
+        from core.llm.providers.anthropic import (
             _API_ALLOWED_KEYS,
         )
